requirements.txt
streamlit
opencv-python-headless
face_recognition
dlib
numpy
pandas
faiss-cpu
